+++ shell.nix
@@ -1,1 +1,7 @@
-(import ./default.nix {}).command.env
+(import ./default.nix {}).project.shellFor {
+  tools = {
+    cabal = "latest";
+    haskell-language-server = "latest";
+    hlint = "latest";
+  };
+}